Enemies now leave a "damage trail" on the floor.
The game was frustrating to play as it was unclear when a tile was safe to walk into or not. Now, tiles will be turned into "goop" while they're lethal, leading to enemies leaving a trail of toxic sludge behind them. Effectively the same mechanics as they had previously, but now with a visual indicator to the player.